requirements.txt
pdfplumber==0.10.3
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0
PyMuPDF>=1.23.0
Pillow>=9.1
